requirements.txt
customtkinter
google-genai
llama-cpp-python
pyttsx3
speechrecognition
platformdirs
requests
openai
